Use float for interest score

+++ footyhints/models/game.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, Text
+from sqlalchemy import Column, Integer, Float, ForeignKey, Text
 from sqlalchemy.orm import relationship
 
 from footyhints.db import session
@@ -20,7 +20,7 @@ class Game(Base):
     home_team = relationship(Team, foreign_keys=[home_team_id], backref='home_games')
     away_team_id = Column(Integer, ForeignKey('teams.id'), nullable=False)
     away_team = relationship(Team, foreign_keys=[away_team_id], backref='away_games')
-    interest_score = Column(Integer, nullable=True)
+    interest_score = Column(Float, nullable=True)
     interest_level = Column(Text, nullable=True)
 
     def __init__(self, **kwargs):
